Yeni kayıtları ReservationData.json'daki kayıtların üzerine eklememe hatası düzetildi, LogRecord class'ında ekleme mi silme mi takibi yapılabilmesi için küçük degisiklik yapıldı

+++ ReservationHandler.cs
@@ -35,16 +35,44 @@ public record class ReservationHandler
         }
         else
         {
-            _reservationRepository.AddReservation(reservation);
-            var logRecord = new LogRecord(ReserverName, reservation.room.roomName, reservation.dateTime);
-            _logHandler.AddLog(logRecord);
+            var _reservations = _reservationRepository.GetAllReservations();
+            bool check = false;
+            foreach (var _reservation in _reservations)
+            {
+                check = false;
+                if (_reservation.dateTime == reservation.dateTime && _reservation.time == reservation.time)
+                {
+                    check = true;//aynısından var
+                    break;
+                }
+            }
+            if (check == true)
+            {
+                Console.WriteLine("Bu saat ve günde başka bir rezervasyon var, lütfen başka saat-gün seçin\n");
+            }
+            else
+            {
+                //ReservationData.jsonda şu anda var olan rezervasyonlar var. Bu yüzden bir rezervasyon silinirse buradan da silinir.
+                _reservationRepository.AddReservation(reservation);
+                DateTime logDateTime = new DateTime(1, 1, reservation.dateTime.Day, reservation.time.Hour, 00, 0);
+                //LogData.jsonda geçmişe yönelik bütün rezervasyon kayıtları var. Bir rezervasyon silinirse buradan silinmez çünkü öyle bir rezervasyon yapıldı önceden.
+                var logRecord = new LogRecord(ReserverName, reservation.room.roomName, logDateTime, "Rezervasyon Eklendi");
+                _logHandler.AddLog(logRecord);
+            }
+
         }
 
 
     }
     public void DeleteReservation(Reservation reservation)
     {
-        _reservationRepository.DeleteReservation(reservation);
+        if (reservation.reserverName!=null && reservation.room.roomName!=null)
+        {
+            _reservationRepository.DeleteReservation(reservation);
+            DateTime logDateTime = new DateTime(1, 1, reservation.dateTime.Day, reservation.time.Hour, 00, 0);
+            var logRecord = new LogRecord(reservation.reserverName, reservation.room.roomName, logDateTime, "Rezervasyon Silindi");
+            _logHandler.AddLog(logRecord);
+        }
     }
     public List<Reservation> GetAllReservations()
     {
